Revert "Newarea notification"
