+++ app/src/main/java/com/example/mobile/commons/LocationManager.kt
@@ -2,9 +2,7 @@ package com.example.mobile.commons
 
 import android.os.Looper
 import androidx.annotation.RequiresPermission
-import androidx.work.PeriodicWorkRequest
 import com.example.mobile.MainActivity
-import com.example.mobile.database.SettingsUtils
 import com.google.android.gms.location.FusedLocationProviderClient
 import com.google.android.gms.location.Granularity
 import com.google.android.gms.location.LocationCallback
@@ -12,9 +10,6 @@ import com.google.android.gms.location.LocationRequest
 import com.google.android.gms.location.LocationResult
 import com.google.android.gms.location.LocationServices
 import com.google.android.gms.location.Priority
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.withContext
-import java.util.concurrent.TimeUnit
 
 object LocationManager {
     private val locationRequest = LocationRequest.Builder(Priority.PRIORITY_HIGH_ACCURACY, 1000 * 60)
@@ -23,7 +18,6 @@ object LocationManager {
         .build()
     private lateinit var fusedLocationClient: FusedLocationProviderClient
     private lateinit var locationCallback: LocationCallback
-    private lateinit var newAreaWorker: PeriodicWorkRequest
 
     //TODO: find sync system so that these values can't be read before they are properly initializeds
     private var _latitude: Double = 0.0
@@ -53,12 +47,6 @@ object LocationManager {
                 }
             }
         }
-
-//        newAreaWorker = PeriodicWorkRequest.Builder(
-//            NewAreaWorker::class.java,
-//            15,
-//            TimeUnit.MINUTES
-//        ).build()
     }
 
     @RequiresPermission(value = "android.permission.ACCESS_FINE_LOCATION")
@@ -72,19 +60,5 @@ object LocationManager {
 
     fun stopLocationRecording() {
         fusedLocationClient.removeLocationUpdates(locationCallback)
-    }
-
-    fun startNotifyingInNewArea() {
-//        withContext(Dispatchers.IO) {
-//            val period = SettingsUtils.storedSettings.
-//        }
-    }
-
-    fun reloadNotificationPeriod() {
-
-    }
-
-    fun stopNotifying() {
-
     }
 }
