+++ app/src/main/java/com/example/mobile/database/MeasurementsUtils.kt
@@ -44,11 +44,6 @@ interface MeasurementDao {
 
     @Query("SELECT COUNT(*) FROM measurement WHERE monitor = :monitor")
     fun countMeasurementsPerMonitor(monitor: MonitorVariant): Int
-
-    @Query("SELECT EXISTS(" +
-           "SELECT * FROM measurement " +
-           "WHERE latitude BETWEEN :bottom AND :top AND longitude BETWEEN :left AND :right)")
-    fun isNewArea(top: Double, bottom: Double, left: Double, right: Double): Boolean
 }
 
 @Entity(tableName = "external_measurement")
@@ -158,9 +153,5 @@ object MeasurementsUtils {
 
     fun getAllExternalMeasurements(variant: MonitorVariant, maxNumber: Int): List<ExternalMeasurement> {
         return DbManager.externalMeasurementDao.getAllExternalMeasurementsPerMonitor(variant, maxNumber)
-    }
-
-    fun isNewArea(top: Double, bottom: Double, left: Double, right: Double): Boolean {
-        return DbManager.measurementDao.isNewArea(top, bottom, left, right)
     }
 }
