+++ app/src/main/java/com/example/mobile/database/SettingsUtils.kt
@@ -1,6 +1,5 @@
 package com.example.mobile.database
 
-import android.text.BoringLayout
 import androidx.room.Dao
 import androidx.room.Entity
 import androidx.room.Insert
@@ -39,7 +38,6 @@ enum class SettingsNames {
     LTE_MEASUREMENT_NUMBER,
     GRID_UNIT_LENGTH,
     ENABLE_PROXIMITY_SHARE,
-    NOTIFY_IN_NEW_AREA,
 }
 
 data class MonitorSettings(
@@ -53,7 +51,6 @@ data class SettingsTable(
     val lte: MonitorSettings = MonitorSettings(1000 * 60, 10),
     val gridUnitLength: Int = 10,
     val enableProximityShare: Boolean = false,
-    val notifyInNewArea: Boolean = false,
 )
 
 object SettingsUtils {
@@ -77,8 +74,7 @@ object SettingsUtils {
                 (map[SettingsNames.LTE_MEASUREMENT_NUMBER.name] ?: "").toInt(),
             ),
             (map[SettingsNames.GRID_UNIT_LENGTH.name] ?: "").toInt(),
-            (map[SettingsNames.ENABLE_PROXIMITY_SHARE.name] ?: "").toBoolean(),
-            (map[SettingsNames.NOTIFY_IN_NEW_AREA.name] ?: "").toBoolean(),
+            (map[SettingsNames.ENABLE_PROXIMITY_SHARE.name] ?: "").toBoolean()
         )
     }
 
@@ -92,7 +88,6 @@ object SettingsUtils {
             Settings(SettingsNames.LTE_MEASUREMENT_NUMBER.name, settings.lte.measurementNumber.toString()),
             Settings(SettingsNames.GRID_UNIT_LENGTH.name, settings.gridUnitLength.toString()),
             Settings(SettingsNames.ENABLE_PROXIMITY_SHARE.name, settings.enableProximityShare.toString()),
-            Settings(SettingsNames.NOTIFY_IN_NEW_AREA.name, settings.notifyInNewArea.toString())
         )
     }
