+++ app/src/main/java/com/example/mobile/screens/SettingsScreen.kt
@@ -1,55 +1,206 @@
 package com.example.mobile.screens
 
+import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.rememberScrollState
-import androidx.compose.foundation.verticalScroll
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.Check
-import androidx.compose.material.icons.filled.Clear
+import androidx.compose.material.icons.filled.ArrowDropDown
+import androidx.compose.material.icons.filled.Info
+import androidx.compose.material.icons.filled.KeyboardArrowDown
+import androidx.compose.material.icons.filled.KeyboardArrowUp
+import androidx.compose.material3.AlertDialog
+import androidx.compose.material3.CircularProgressIndicator
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
 import androidx.compose.material3.LinearProgressIndicator
 import androidx.compose.material3.OutlinedButton
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextField
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
 import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.text.input.KeyboardType
+import androidx.compose.ui.unit.dp
 import com.example.mobile.R
-import com.example.mobile.commons.LocationManager
 import com.example.mobile.composables.AlertSeverity
 import com.example.mobile.composables.AlertTextbox
-import com.example.mobile.composables.NumberSetting
-import com.example.mobile.composables.OptionsSetting
-import com.example.mobile.composables.SettingLayout
-import com.example.mobile.composables.SwitchSetting
+import com.example.mobile.composables.OptionSelect
 import com.example.mobile.database.MonitorSettings
-import com.example.mobile.database.SettingsNames
 import com.example.mobile.database.SettingsTable
 import com.example.mobile.database.SettingsUtils
 import com.example.mobile.monitors.MonitorVariant
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.withContext
 
+@Composable
+fun SettingLayout(
+    title: String,
+    description: String = "",
+    inputField: @Composable() () -> Unit
+) {
+    var showDialog: Boolean by remember { mutableStateOf(false) }
+
+    Column(
+        modifier = Modifier
+            .padding(16.dp)
+    ) {
+        Row(
+            horizontalArrangement = Arrangement.Center,
+            verticalAlignment = Alignment.CenterVertically
+        ) {
+            Text(
+                text = title,
+                fontWeight = FontWeight.Bold,
+                modifier = Modifier
+                    .padding(start = 8.dp)
+            )
+            Spacer(modifier = Modifier.weight(1f))
+            if (description != "") {
+                IconButton(onClick = {
+                    showDialog = true
+                }) {
+                    Icon(
+                        imageVector = Icons.Default.Info,
+                        contentDescription = "Setting explanation"
+                    )
+                }
+            }
+
+            if (showDialog) {
+                AlertDialog(
+                    title = {
+                        Text(text = title)
+                    },
+                    text = {
+                        Text(text = description)
+                    },
+                    onDismissRequest = {
+                        showDialog = false
+                    },
+                    confirmButton = { },
+                    dismissButton = { }
+                )
+
+            }
+        }
+        inputField()
+    }
+}
+
+@Composable
+fun NumberSetting(
+    title: String,
+    description: String,
+    value: Long,
+    onChange: (newValue: Long) -> Unit,
+    max: Int = Int.MAX_VALUE,
+    min: Int = Int.MIN_VALUE
+) {
+    SettingLayout(
+        title = title,
+        description = description,
+        inputField = {
+            Row(
+                modifier = Modifier
+                    .fillMaxWidth(),
+                horizontalArrangement = Arrangement.Center,
+                verticalAlignment = Alignment.CenterVertically
+            ) {
+                TextField(
+                    value = value.toString(),
+                    onValueChange = {
+                        val newValue = it.toLongOrNull() ?: 0
+                        onChange(newValue)
+                    },
+                    keyboardOptions = KeyboardOptions.Default.copy(
+                        keyboardType = KeyboardType.Number
+                    ),
+                    leadingIcon = {
+                        IconButton(
+                            onClick = {
+                                if (value > min) {
+                                    onChange(value - 1)
+                                }
+                            },
+                            modifier = Modifier
+                                .size(48.dp)
+                                .padding(8.dp)
+                        ) {
+                            Icon(imageVector = Icons.Default.KeyboardArrowDown, contentDescription = "Decrease")
+                        }
+                    },
+                    trailingIcon = {
+                        IconButton(
+                            onClick = {
+                                if (value < max) {
+                                    onChange(value + 1)
+                                }
+                            },
+                            modifier = Modifier
+                                .size(48.dp)
+                                .padding(8.dp)
+                        ) {
+                            Icon(imageVector = Icons.Default.KeyboardArrowUp, contentDescription = "Increase")
+                        }
+                    },
+                    modifier = Modifier
+                        .weight(1f)
+                        .fillMaxWidth()
+                )
+            }
+        }
+    )
+}
+
+@Composable
+fun <T> OptionsSetting(
+    title: String,
+    description: String,
+    value: T,
+    options: List<T>,
+    onChange: (newValue: T) -> Unit,
+    getLabel: (option: T) -> String = { it.toString() },
+) {
+    SettingLayout(
+        title = title,
+        description = description,
+        inputField = {
+            OptionSelect(
+                label = "",
+                options = options,
+                onChange = {
+                    onChange(it)
+                },
+                value = value,
+                getLabel = getLabel,
+                buttonModifier = Modifier.fillMaxWidth()
+            )
+        }
+    )
+}
 
 @Composable
 fun SettingsScreen(
     variant: MonitorVariant,
     navigateTo: (targetScreen: Screens) -> Unit,
-    startNotifyingInNewArea: () -> Unit,
-    stopNotifying: () -> Unit,
 ) {
     val context = LocalContext.current
-    val scrollState = rememberScrollState()
-    //TODO: properly initialize
     val gridSizes = listOf(10, 100, 1000)
 
     var settings: SettingsTable? by remember { mutableStateOf(null) }
     var monitorSettings: MonitorSettings? by remember { mutableStateOf(null) }
-    var notifyInNewArea by remember { mutableStateOf(false) }
     var initializing by remember { mutableStateOf(true) }
 
     fun setMonitorSettings() {
@@ -59,13 +210,6 @@ fun SettingsScreen(
             MonitorVariant.LTE -> settings!!.lte
         }
     }
-
-//    fun startNotifyingInNewArea() {
-//        LocationManager.startNotifyingInNewArea()
-//    }
-//    fun stopNotifying() {
-//        LocationManager.stopNotifying()
-//    }
 
     LaunchedEffect(Unit) {
         withContext(Dispatchers.IO) {
@@ -86,37 +230,13 @@ fun SettingsScreen(
         if(!initializing) setMonitorSettings()
     }
 
-    var skipFirst by remember { mutableStateOf(true) }
-    LaunchedEffect(notifyInNewArea) {
-        if (notifyInNewArea) {
-            startNotifyingInNewArea()
-        }
-        else {
-            stopNotifying()
-        }
-        // on the first render we shouldn't override the setting with false
-        if (skipFirst) {
-            skipFirst = false
-        }
-        else {
-            withContext(Dispatchers.IO) {
-                SettingsUtils.updateSingleSetting(
-                    SettingsNames.NOTIFY_IN_NEW_AREA,
-                    notifyInNewArea.toString()
-                )
-            }
-        }
-    }
-
     if (initializing)
         LinearProgressIndicator(modifier = Modifier.fillMaxWidth())
-    else Column(
-        modifier = Modifier.verticalScroll(scrollState)
-    ) {
-        AlertTextbox(
-            severity = AlertSeverity.INFO,
-            content = LocalContext.current.getString(R.string.settings_alert_note)
-        )
+    else Column {
+//        AlertTextbox(
+//            severity = AlertSeverity.INFO,
+//            content = LocalContext.current.getString(R.string.settings_alert_note)
+//        )
 
         NumberSetting(
             title = context.getString(R.string.period_setting_title),
@@ -185,14 +305,6 @@ fun SettingsScreen(
             },
             options = gridSizes
         )
-        SwitchSetting(
-            label = "label",
-            onClick = {
-                notifyInNewArea = !notifyInNewArea
-            },
-            value = notifyInNewArea,
-        )
-
         SettingLayout(
             title = context.getString(R.string.settingscreen_export_button_description),
             inputField = {
